Fixed some style errors on the web pages.
[Fixes issue 7](https://github.com/camilosampedro/Aton/issues/7)

1. Why this change is necessary?
It makes the general style of the application feel more comfortable and
makes the left menu to feel like a menu.

2. How does it address the issue?
`.content` now has a height of `calc(100vh - 45px)` which 45px is the
`header` height. It also has a `overflow-y: scroll;` to avoid
overfilling on y axis.

3. What side effects does this change have?
It only changes visual styling (Frontend styling).