Panels' padding and margins fixed. SSH Functions comments added.
1. Why this change is necessary?
It helps to view the main webs better. Also comments help to create a comment
syntax for SSH Orders on SSHFunctions class.

2. How does it address the issue?
Adding a padding a margin rule to .panel and .room-panel CSS classes.
Adding a ScalaDoc comment for every single SSH order in SSHFunctions class.

3. What side effects does this change have?
It should improve user interface but it could change some other web interfaces,
it has to be tested.