Initial translation to German (Fixes #22)
This commit adds a first translation of the
Spanish strings to German. This is necessary
to allow German users who are not fluent in
Spanish to use the application too.

It adds a messages.de file to the conf dir-
rectory with all currently known messages
translated to German. Most messages are
pretty straight-forward, but for some,
multiple interpretations are possible.
These might be revised in a future
update.