Fix German suggestion translations
This commit adjusts the German translation to better
reflect what suggestions are for. This is necessary
to help users better understand that feature.
Previously, it was labeled with the literal
translation of suggestions, which turned out to
be inappropriate for this usage. Now, it uses the
literal translation of request, which describes
the feature better. This change does not have
any known caveats.