Add missing translation strings for login page
This commit adds two more translation endpoints on the login page,
the caption of the login button and the placeholder of the password
field. It also adds translations to the English and Spanish messages.

This change is necessary because not all languages have the same
words for password and login.

This commit addresses the issue by changing the login template
to request translations for these two strings and adding
respective translations to the existing messages files.

This change adds new strings, so all translations that are not
shipped need to update accordingly.