Fix Laboratory list i18n
This commit makes use of the previously unused
laboratory.list.title string and also uses the delete and edit
strings for each laboratory in the list, instead of defaulting
to static strings in Spanish. This is necessary to improve
translation coverage. The implementation is trivial. The only
known issue with this is that new translations might not have
that string because they thought it was not implemented, but
that could be considered a minor caveat.