Fix missed suggestion translation
This commit fixes the menu name for suggestions in German
being different from the word used everywhere else. The
reason it was different is because the menu items
are declared somewhere else than the rest of the
suggestion-related strings, so I missed it when updating
the translation with the new word. This is necessary
because we strive to be consistent in our wording.
The implementation is trivial, no cavets known.