Action States added (#36)
1. Why this change is necessary?
For adding a better integration between dao to services and services to controllers

2. How does it address the issue?
There was added `ActionState` objects for identifying the states of DAO and Services operations. There were also removed all kind of DAOs from controllers, replacing them with services.

* Action States added to ComputerController

* Fixed states and tests

* All controllers now work with ActionStates

* UserDAO replaced with UserService in controllers

* Added ExecutionContext

* Added UserService

3. What side effects does this change have?
It should have a little impact on the functional Aton operation. The more risky thing I expect is to don't have well recognized the return of Slick's `delete` on the DAO layer.