Angular2 transpilation completely fixed
Solves [issue 41](https://github.com/camilosampedro/Aton/issues/41)
1. Why this change is necessary?
There was a lot of troubles on adding the Angular2 traspilation to Aton. This issue solves them,
it is now necessary to program the application with Angular2.

2. How does it address the issue?
Angular2 app reseted to Heroes' Angular2 default example

3. What side effects does this change have?
It is a big change on the client-side. Nothing in Angular2 is actually working, it is just
now integrated into Play lifecycle