MySQL database migrated to H2
Solves **[issue 50](https://github.com/camilosampedro/Aton/issues/50)**
1. Why this change is necessary?
It is necessary to remove the MySQL dependency to run Aton, to be a more independent application.

2. How does it address the issue?
Changing database source from MySQL to H2, it was necessary to change the creation script and all table
mappings to uppercase. Before running the application it is needed to check if the database has been
previously initialized (Tables, indexes and initial values) with DatabaseInitializer.

3. What side effects does this change have?
It could be a big change to the application, this still has to be tested.