Salting to passwords added.
This solves **[issue 6](https://github.com/camilosampedro/Aton/issues/6)**
1. Why this change is necessary?
Aton's passwords inside its database was open and raw to anyone's look. Those passwords are not necessary to keep raw, just
for username-password checking.

2. How does it address the issue?
Using JBCrypt library for hashing and checking.

3. What side effects does this change have?
It is not a big change, though it could be necessary to test user's log-in.